Tidy up keyword arguments to Schema::Value
Now seamlessly handles old style struct declarations, while using
keyword_init to construct the bare struct.

Only the public interface is set up to support passing a list of
attributes to the factory and instance constructor. List handling and
any-type mapping is removed from the validator so that things are more
internally consistent.