Add raw type abstraction layer to schema system
This just covers some basics that are needed initially, but it shouldn’t
be too hard to expand this to cover validation and type coercion now
that this abstraction is in place.

Also introduces an initial stab at a Markdown/HTML format type that can
be used for content fields.